Got flash working and enabled config.hpp:
The config was updated to have logging enabled

kalman.hpp:

The kalman filter had a const defined function for `.state()`. This
became a bit messy with the LogMessage struct. removing the const
'fixed' the issue.

flash.cpp:

The flash chip we are using is different from the previous flight
computer. Previously a W25Q16 and we are using the W25Q32. The
device instructions, at least the ones we use, are the same for both
devices. W25Q32 had a different device id and for the most part was the
only change.

main.cpp:

First change was to comment out the "Unrecognized command." output.
A bit annoying and spams the serial terminal a bit. Can be uncommented
if needed.

Next was uncommenting the `log_start()`. Very important in being able to
use the flash chip for logging the flight data.

Final change was setting sys_v to 0 as our board can not yet get this
information.